- Refactor CLI commands
