+++ spec/approvals/cli/commands
@@ -1,0 +1,8 @@
+Completely - Bash Completions Generator
+
+Commands:
+  init      Create a new sample YAML configuration file
+  preview   Generate the bash completion script to STDOUT
+  generate  Generate the bash completion script to a file
+
+Run completely COMMAND --help for more information
