+++ spec/approvals/cli/generate/custom-out-path
