+++ spec/approvals/cli/generate/custom-path
