+++ spec/approvals/cli/generate/function
