+++ spec/approvals/cli/generate/help
@@ -1,0 +1,23 @@
+Generate the bash completion script to a file
+
+Usage:
+  completely generate [CONFIG_PATH OUTPUT_PATH --function NAME --wrap NAME]
+  completely generate (-h|--help)
+
+Options:
+  -f --function NAME
+    Modify the name of the function in the generated script
+
+  -w --wrap NAME
+    Wrap the completion script inside a function that echos the script. This is
+    useful if you wish to embed it directly in your script
+
+  -h --help
+    Show this help
+
+Parameters:
+  CONFIG_PATH
+    Path to the YAML configuration file [default: completely.yaml]
+
+  OUTPUT_PATH
+    Path to the output bash script [default: completely.bash]
