+++ spec/approvals/cli/generate/no-args
