+++ spec/approvals/cli/generate/wrapper
@@ -1,0 +1,1 @@
+Saved completely.bash
