--- spec/approvals/cli/help
@@ -1,47 +1,0 @@
-Bash completion script generator
-
-Usage:
-  completely new [CONFIG_PATH]
-  completely preview [CONFIG_PATH --function NAME]
-  completely generate [CONFIG_PATH OUTPUT_PATH --function NAME --wrap NAME]
-  completely (-h|--help|--version)
-
-Commands:
-  new
-    Create a new sample YAML configuration file
-
-  preview
-    Generate the bash completion script to STDOUT
-
-  generate
-    Generate the bash completion script to a file
-
-Options:
-  -f --function NAME
-    Modify the name of the function in the generated script
-
-  -w --wrap NAME
-    Wrap the completion script inside a function that echos the script. This is
-    useful if you wish to embed it directly in your script
-
-  -h --help
-    Show this help
-
-  --version
-    Show version number
-
-Parameters:
-  CONFIG_PATH
-    Path to the YAML configuration file
-
-  OUTPUT_PATH
-    Path to the output bash script
-
-Examples:
-  completely new
-  completely new input.yaml
-  completely preview --function _my_completions
-  completely generate
-  completely generate input.yaml
-  completely generate input.yaml output.sh -f _my_completions
-  completely generate -w give_comps -f my_completions
