+++ spec/approvals/cli/init/no-args
